Re-arrange the repository move the its test to itsmft directory and re-arrange the structure
