+++ itsmft/CMakeLists.txt
@@ -17,5 +17,6 @@
 add_subdirectory(common)
 add_subdirectory(its)
 add_subdirectory(mft)
+add_subdirectory(test)
 
 message(STATUS "itsmft enabled")
